+++ forui/lib/src/widgets/accordion/accordion_item.dart
@@ -42,6 +42,12 @@ class FAccordionItem extends StatefulWidget with FAccordionItemMixin {
   /// {@macro forui.foundation.doc_templates.onFocusChange}
   final ValueChanged<bool>? onFocusChange;
 
+  /// {@macro forui.foundation.FTappable.onHoverChange}
+  final ValueChanged<bool>? onHoverChange;
+
+  /// {@macro forui.foundation.FTappable.onStateChange}
+  final ValueChanged<Set<WidgetState>>? onStateChange;
+
   /// The child.
   final Widget child;
 
@@ -55,6 +61,8 @@ class FAccordionItem extends StatefulWidget with FAccordionItemMixin {
     this.autofocus = false,
     this.focusNode,
     this.onFocusChange,
+    this.onHoverChange,
+    this.onStateChange,
     super.key,
   });
 
@@ -69,7 +77,9 @@ class FAccordionItem extends StatefulWidget with FAccordionItemMixin {
       ..add(FlagProperty('initiallyExpanded', value: initiallyExpanded, ifTrue: 'Initially expanded'))
       ..add(FlagProperty('autofocus', value: autofocus, defaultValue: false, ifTrue: 'autofocus'))
       ..add(DiagnosticsProperty('focusNode', focusNode))
-      ..add(ObjectFlagProperty.has('onFocusChange', onFocusChange));
+      ..add(ObjectFlagProperty.has('onFocusChange', onFocusChange))
+      ..add(ObjectFlagProperty.has('onHoverChange', onHoverChange))
+      ..add(ObjectFlagProperty.has('onStateChange', onStateChange));
   }
 }
 
@@ -122,6 +132,8 @@ class _FAccordionItemState extends State<FAccordionItem> with TickerProviderStat
           autofocus: widget.autofocus,
           focusNode: widget.focusNode,
           onFocusChange: widget.onFocusChange,
+          onHoverChange: widget.onHoverChange,
+          onStateChange: widget.onStateChange,
           onPress: () => controller.toggle(index),
           builder: (_, states, _) => Padding(
             padding: style.titlePadding,